feat(gateway): add DNS lookup duration histogram metric
The Gateway resolves DNS-resource domains against its upstream resolver
but, unlike the Client, did not record how long this takes. Add the same
`dns.lookup.duration` histogram so resolution latency (including
timeouts) is observable on Gateways too.

The metric definition is hoisted into the shared `telemetry` crate so
both the Client and Gateway emit it under the `connlib` meter.
