+++ rust/libs/connlib/tunnel/src/otel.rs
@@ -66,21 +66,6 @@ pub mod attr {
 
 pub mod metrics {
     pub use telemetry::otel::metrics::*;
-
-    use opentelemetry::metrics::Histogram;
-
-    /// Measures how long connlib takes to recursively resolve a DNS query against
-    /// an upstream resolver (locally or through the tunnel).
-    pub fn dns_lookup_duration() -> Histogram<f64> {
-        opentelemetry::global::meter("connlib")
-            .f64_histogram("dns.lookup.duration")
-            .with_description("Duration of a recursive DNS lookup against an upstream resolver.")
-            .with_unit("s")
-            .with_boundaries(vec![
-                0.005, 0.01, 0.025, 0.05, 0.075, 0.1, 0.25, 0.5, 0.75, 1.0, 2.5, 5.0, 7.5, 10.0,
-            ])
-            .build()
-    }
 }
 
 #[cfg(test)]
