+++ rust/libs/telemetry/src/otel.rs
@@ -478,7 +478,7 @@ pub mod metrics {
 
     use opentelemetry::{
         KeyValue,
-        metrics::{Counter, Gauge},
+        metrics::{Counter, Gauge, Histogram},
     };
 
     use crate::otel::QueueLength;
@@ -504,6 +504,18 @@ pub mod metrics {
             .u64_gauge("tunnel.connection.count")
             .with_description("Number of connections by the network path in use.")
             .with_unit("{connection}")
+            .build()
+    }
+
+    /// Measures how long connlib takes to recursively resolve a DNS query against an upstream resolver.
+    pub fn dns_lookup_duration() -> Histogram<f64> {
+        opentelemetry::global::meter("connlib")
+            .f64_histogram("dns.lookup.duration")
+            .with_description("Duration of a recursive DNS lookup against an upstream resolver.")
+            .with_unit("s")
+            .with_boundaries(vec![
+                0.005, 0.01, 0.025, 0.05, 0.075, 0.1, 0.25, 0.5, 0.75, 1.0, 2.5, 5.0, 7.5, 10.0,
+            ])
             .build()
     }
